Moved inside Quake folder
Moved the entire project inside the Quake folder so that all the data
inside is accessible from the app. The app mostly lives inside the
‘QuakeWebTools’ folder, with the exception of the ‘QuakeWebTools.html’
file, which needs to be in the Quake root in order to gain access to
sub-folders, such as ID1.

Note: on MacOSX and Linux, Quake folders and filenames must be made
lower-case for anything to work. However, the files and folder names
are upper-case on Windows, so this app may not work correctly there.